3_crew/video_generation_crew/pyproject.toml
[project]
name = "video_generation_crew"
version = "0.1.0"
description = "video_generation_crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.152.0,<1.0.0",
    "google-genai>=1.29.0",
    "python-dotenv>=1.0.0,<2.0.0",
]

[project.scripts]
video_generation_crew = "video_generation_crew.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
